add more tests

+++ src/BarkerMCMC.jl
@@ -40,7 +40,7 @@ Livingstone, S., Zanella, G., 2021. The Barker proposal: Combining robustness an
 function barker_mcmc(log_p::Function, ∇log_p::Function,
                      inits::AbstractVector;
                      n_iter = 100::Int, σ = 2.4/(length(inits)^(1/6)),
-                     target_acceptance_rate = 0.4, κ = 0.6,
+                     target_acceptance_rate = 0.4, κ::Float64 = 0.6,
                      n_iter_adaptation = Inf,
                      preconditioning::Function = precond_eigen)
